requirements.txt

rembg
numpy
Pillow
Flask
